requirements.txt
# Common libs
djangorestframework>=3.11.0
PyJWT==1.7.1
sc4py==0.1.1
sc4net==0.1.2

# Development
django_debug_toolbar==2.2
django-extensions==2.2.8


# Build
twine==1.12.1
flake8==3.7.9
coverage==5.0.3
